pyproject.toml
[project]
name = "hermes-agent"
version = "0.1.0"
description = "HermesAgent (Taskspine) — strict orchestration harness for low→plan→execute agent stacks"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.15",
  "rich>=13.9",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "huggingface-hub>=0.25",
  "openai>=1.55",
]

[project.scripts]
hermes = "hermes_agent.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
